# refactor: Apply new storage