--- a/src/new_storage/backend/mod.rs
+++ b/src/new_storage/backend/mod.rs
@@ -6,7 +6,7 @@ pub mod memory_backend;
 use std::fmt::Debug;
 
 use async_trait::async_trait;
-pub use backend_impl::BackendImpl;
+pub use backend_impl::{BackendBuilder, BackendImpl};
 
 use super::error::StorageResult;