--- a/src/new_storage/block.rs
+++ b/src/new_storage/block.rs
@@ -1,14 +1,41 @@
 //! The definition of the `Block`.
 
+use std::fmt::Debug;
+
+use clippy_utilities::OverflowArithmetic;
+
 /// The size of a block in bytes (for test).
 #[cfg(test)]
 pub const BLOCK_SIZE: usize = 4 * 1024 * 1024;
+
+/// Make a partial slice of `u8` slice for debugging.
+fn make_partial_u8_slice(src: &[u8], start: usize, len: usize) -> String {
+    let end = start.overflow_add(len).min(src.len());
+    let show_length = end.overflow_sub(start);
+    let slice = src.get(start..end).unwrap_or(&[]);
+
+    let mut result = String::from("[");
+
+    for (i, &ch) in slice.iter().enumerate() {
+        result.push_str(format!("{ch}").as_str());
+        if i != show_length.overflow_sub(1) {
+            result.push_str(", ");
+        }
+    }
+
+    if end != src.len() {
+        result.push_str(", ...");
+    }
+
+    result.push(']');
+
+    result
+}
 
 /// Represents a block of data.
 ///
 /// A `Block` contains a vector of bytes representing the block data,
 /// along with metadata such as pin count, dirty flag, and version.
-#[derive(Debug)]
 pub struct Block {
     /// The underlying data
     data: Vec<u8>,
@@ -18,6 +45,20 @@ pub struct Block {
     dirty: bool,
     /// The version of the block
     version: usize,
+}
+
+impl Debug for Block {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("Block")
+            .field(
+                "data",
+                &make_partial_u8_slice(&self.data, 0, self.data.len()),
+            )
+            .field("pin_count", &self.pin_count)
+            .field("dirty", &self.dirty)
+            .field("version", &self.version)
+            .finish()
+    }
 }
 
 impl Block {