Since this PR, we will handle all storage requests with the new storage.