@Reto911 2024-05-02
The new storage module has better performance on write while has worse performance on read, compared with the old one. The result of a performance test are below.

#### Write
block size (MiB) | new (MiB/s) | old (MiB/s)
-- | -- | --
0.5 | 36.4 | 30.8
1 | 33.4 | 28.1
2 | 33 | 20.2
4 | 28.8 | 14.6

#### Read
block size (MiB) | new (MiB/s) | old (MiB/s)
-- | -- | --
0.5 | 205 | 502
1 | 238 | 517
2 | 260 | 530
4 | 289 | 466

#### Read (4 jobs)
block size (MiB) | new (MiB/s) | old (MiB/s)
-- | -- | --
0.5 | 814 | 2435
1 | 871 | 2207
2 | 901 | 2634
4 | 906 | 2688

#### Read (8 jobs)
block size (MiB) | new (MiB/s) | old (MiB/s)
-- | -- | --
0.5 | 1104 | 5038
1 | 1163 | 5107
2 | 1235 | 5224
4 | 1217 | 5261